CONTINGENT2-831 Разработка триггера формирования топика AreaInfo для участков с МР с окончившимся сроком действия

+++ contingent/application/contingent-area-domain/src/main/java/moscow/ptnl/contingent/domain/area/repository/AreaRepository.java
@@ -41,5 +41,5 @@ public interface AreaRepository {
 
     Area save(Area area);
 
-    List<Area> findAreasForSyncToK1(int daysForSelect);
+    List<Area> findAreasForSyncToK1(long daysForSelect);
 }
